moves the mockDSData.js file into a test directory inside the data directory
